23-06-29 단순 수정

+++ _mina/code/Integrate_training_prep.py
@@ -1,15 +1,21 @@
 import cv2
 import os
 import shutil
+import platform
 import yaml
 import json
 import deeplabcut
 import pandas as pd
 from tqdm import tqdm
+"""
+    Dog Pose Estimation - Sensor Train :  AI-Hub 데이터셋 활용하여 데이터 전처리 & DLC 학습시키기
+    
+    main() 함수를 실행하면, AI-Hub 데이터셋을 활용하여 DeepLabCut을 학습시키는 과정을 진행한다.
+"""
 
 
 
-def frames_to_video(src_path, dataset):
+def frames_to_video(src_path, dataset, pbar):
     """
         Convert frames to AVI video.
 
@@ -20,32 +26,46 @@ def frames_to_video(src_path, dataset):
         
         dataset : list
             List of the dataset.
+        
+        pbar : tqdm
+            Progress bar object.
     """
     for i in range(len(src_path)):
         for j in range(len(dataset[i])):
             frames_folder = src_path[i] + '/images/' + dataset[i][j]          # dataset의 frame들의 폴더 위치
             video_name = src_path[i] + '/videos/' + dataset[i][j] + '.avi'           # 비디오 저장위치
+            
+            if not os.path.exists(os.path.join(src_path[i], 'videos')):
+                os.makedirs(os.path.join(src_path[i], 'videos'))
 
-            # 리스트 내포(list comprehension) => images = []    for img in os.listdir(frames_folder):   if img.endswith(".jpg"):   images.append(img)
-            images = [img for img in os.listdir(frames_folder) if img.endswith(".jpg")]
-            # 파일 이름을 기준으로 정렬하여 목록 생성
-            images.sort(key=lambda x: int(x.split('_')[1]))
-            frame = cv2.imread(os.path.join(frames_folder, images[0]))
-            height, width, layers = frame.shape
+            if not os.path.exists(video_name):
 
-            # 세 번째 인수가 FPS. 높을 수록 부드럽게 재생
-            video = cv2.VideoWriter(video_name, 0, 6, (width,height))
+                images = [img for img in os.listdir(frames_folder) if img.endswith(".jpg")]
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
-            for image in images:
-                video.write(cv2.imread(os.path.join(frames_folder, image)))
+                # 파일 이름을 기준으로 정렬하여 목록 생성
+                images.sort(key=lambda x: int(x.split('_')[1]))
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
+                
+                frame = cv2.imread(os.path.join(frames_folder, images[0]))
+                height, width, layers = frame.shape
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
+                
+                # 세 번째 인수가 FPS. 높을 수록 부드럽게 재생
+                video = cv2.VideoWriter(video_name, 0, 6, (width,height))
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
-            cv2.destroyAllWindows()
-            video.release()
+                for image in images:
+                    video.write(cv2.imread(os.path.join(frames_folder, image)))
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
+
+                cv2.destroyAllWindows()
+                video.release()
 
 
 
 ## https://deeplabcut.github.io/DeepLabCut/docs/standardDeepLabCut_UserGuide.html#a-create-a-new-project
-def create_new(project, scorer, working_directory, src_path, dataset):
+def create_new(project, scorer, working_directory, src_path, dataset, pbar):
     """
         Create a new project for pose estimation.
 
@@ -65,6 +85,9 @@ def create_new(project, scorer, working_directory, src_path, dataset):
 
         dataset : list
             List of the action of dataset.
+        
+        pbar : tqdm
+            Progress bar object.
 
         Returns
         -------
@@ -76,29 +99,43 @@ def create_new(project, scorer, working_directory, src_path, dataset):
     """
     vdos = []       # video set 
     for i in range(len(src_path)):
+        pbar.update((1 / len(src_path)) * 0.2)
+
         if not os.path.exists(src_path[i] + '/videos'):
             os.makedirs(src_path[i] + '/videos')
+        pbar.update((1 / len(src_path)) * 0.2)
+
         for j in range(len(dataset[i])):
             full_path_of_video = src_path[i] + '/videos/' + dataset[i][j] + '.avi'
-            inputPath = full_path_of_video.replace('/', '\\')
-            vdos.append(inputPath)
+            if platform.system() == 'Windows':
+                inputPath = full_path_of_video.replace('/', '\\')
+                vdos.append(inputPath)
+            else :
+                vdos.append(full_path_of_video)
+        pbar.update((1 / len(src_path)) * 0.3)
+
     config_path = deeplabcut.create_new_project(project, scorer, vdos, working_directory=working_directory, copy_videos=True)         
+    pbar.update(1 - (0.2 + 0.2 + 0.3))
     return config_path, vdos
 
 
 
 ## https://github.com/DeepLabCut/DeepLabCut/blob/main/deeplabcut/create_project/new.py#L21
-def config_edit(config_path):
+def config_edit(config_path, pbar):
     """
         Edit the config.yaml file to overwrite the bodyparts and skeleton, etc.
 
         Parameters
         ----------
         config_path : string
-            Full path of the config.yaml file as a string.
+            Full path of the config.yaml file as a string.   
+            
+        pbar : tqdm
+            Progress bar object.
     """
     with open(config_path, 'r') as file:
             cfg_file = yaml.safe_load(file)
+    pbar.update(0.5)
 
     cfg_file['bodyparts'] = ['Nose', 'Forehead', 'MouthCorner', 'LowerLip', 'Neck', 'RightArm', 'LeftArm', 'RightWrist', 'LeftWrist', 'RightFemur', 'LeftFemur', 'RightAnkle', 'LeftAnkle', 'TailStart', 'TailTip']
     cfg_file["skeleton"] = [
@@ -122,11 +159,12 @@ def config_edit(config_path):
 
     with open(config_path, 'w') as file:
         yaml.dump(cfg_file, file)
+    pbar.update(0.5)
 
 
 
 ## https://github.com/DeepLabCut/DeepLabCut/wiki/Using-labeled-data-in-DeepLabCut-that-was-annotated-elsewhere/35eb6bc2079d3e1125dafa87b05e07e47df388d3
-def json_to_csv(scorer, src_path, lab_path, conclusion_path, dataset):
+def json_to_csv(scorer, src_path, lab_path, conclusion_path, dataset, pbar):
     """
         Extract joint coordinates in a JSON file and convert to a CSV file.
 
@@ -146,66 +184,78 @@ def json_to_csv(scorer, src_path, lab_path, conclusion_path, dataset):
 
         dataset : list
             List of the dataset.
+            
+        pbar : tqdm
+            Progress bar object.
     """
     for i in range(len(src_path)):
         for j in range(len(dataset[i])):
             json_path = lab_path[i] + '/json/' + dataset[i][j] + '.json'             # coords가 있는 json 경로
-            if not os.path.exists(lab_path[i] + '/csv/'):
-                os.makedirs(lab_path[i] + '/csv/')
             csv_path = lab_path[i] + '/csv/' + dataset[i][j] + '.csv'            # 저장될 csv 경로
             frames_folder = src_path[i] + '/images/' + dataset[i][j]             # frames 폴더 경로
             labeled_path = 'labeled-data/' + dataset[i][j] + '/'              # ImgPrefixPath
             copy_csv_path = conclusion_path + '/labeled-data/' + dataset[i][j] + '/CollectedData_' + scorer + '.csv'            # copied csv save 경로
             copy_img_src = frames_folder                # copy 할 framses directory 경로
             copy_img_dst = conclusion_path + '/labeled-data/' + dataset[i][j]           # frames copy and save 경로
+            pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
+            
+            if not os.path.exists(lab_path[i] + '/csv/'):
+                os.makedirs(lab_path[i] + '/csv/')
+            
+            if not os.path.exists(csv_path) or not os.path.exists(copy_csv_path):
+                  
+                with open(json_path, 'r', encoding='utf-8') as f:
+                    data = json.load(f)
 
-            with open(json_path, 'r', encoding='utf-8') as f:
-                data = json.load(f)
+                # 선택한 키의 값만 추출
+                selected_data = []
+                for annotation in data['annotations']:
+                    row = []
+                    for keypoint in annotation['keypoints'].values():
+                        if keypoint is not None:
+                            row.extend([float(keypoint['x']), float(keypoint['y'])])
+                        else:
+                            row.extend([None, None])
+                    selected_data.append(row)
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
-            # 선택한 키의 값만 추출
-            selected_data = []
-            for annotation in data['annotations']:
-                row = []
-                for keypoint in annotation['keypoints'].values():
-                    if keypoint is not None:
-                        row.extend([float(keypoint['x']), float(keypoint['y'])])
-                    else:
-                        row.extend([None, None])
-                selected_data.append(row)
+                # Create a DataFrame with the selected data and column names
+                df = pd.DataFrame(selected_data, columns=[scorer] * len(selected_data[0]))
 
-            # Create a DataFrame with the selected data and column names
-            df = pd.DataFrame(selected_data, columns=[scorer] * len(selected_data[0]))
+                df.loc[-1] = ['x', 'y'] * ((len(df.columns)) // 2)
+                df.index = df.index + 1
+                df = df.sort_index()
+            
+                # 리스트 내포(list comprehension)를 사용하여 각 요소가 두 번씩 반복되는 리스트 생성
+                bodyparts = ['Nose', 'Forehead', 'MouthCorner', 'LowerLip', 'Neck', 'RightArm', 'LeftArm', 'RightWrist', 'LeftWrist', 
+                            'RightFemur', 'LeftFemur', 'RightAnkle', 'LeftAnkle', 'TailStart', 'TailTip']
+                df.loc[-1] = [part for part in bodyparts for _ in range(2)]
+                df.index = df.index + 1
+                df = df.sort_index()
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
-            df.loc[-1] = ['x', 'y'] * ((len(df.columns)) // 2)
-            df.index = df.index + 1
-            df = df.sort_index()
-        
-            # 리스트 내포(list comprehension)를 사용하여 각 요소가 두 번씩 반복되는 리스트 생성
-            bodyparts = ['Nose', 'Forehead', 'MouthCorner', 'LowerLip', 'Neck', 'RightArm', 'LeftArm', 'RightWrist', 'LeftWrist', 
-                         'RightFemur', 'LeftFemur', 'RightAnkle', 'LeftAnkle', 'TailStart', 'TailTip']
-            df.loc[-1] = [part for part in bodyparts for _ in range(2)]
-            df.index = df.index + 1
-            df = df.sort_index()
+                images = [img for img in os.listdir(frames_folder) if img.endswith(".jpg")]
+                images.sort(key=lambda x: int(x.split('_')[1]))
+                project_dataPath = labeled_path             # 앞에 labeled-data path 추가                  
+                images = [project_dataPath + img for img in images]
+                df.insert(0, 'scorer', ['bodyparts', 'coords'] + images)            # 0번 째 열에 삽입
 
-            images = [img for img in os.listdir(frames_folder) if img.endswith(".jpg")]
-            images.sort(key=lambda x: int(x.split('_')[1]))
-            project_dataPath = labeled_path             # 앞에 labeled-data path 추가                  
-            images = [project_dataPath + img for img in images]
-            df.insert(0, 'scorer', ['bodyparts', 'coords'] + images)            # 0번 째 열에 삽입
+                # Save the resulting DataFrame to a CSV file
+                df.to_csv(csv_path, index=False, header=True)
+                shutil.copy(csv_path, copy_csv_path)
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
-            # Save the resulting DataFrame to a CSV file
-            df.to_csv(csv_path, index=False, header=True)
-            shutil.copy(csv_path, copy_csv_path)
-            # Copy the sourceFrames 
-            for file_name in os.listdir(copy_img_src):
-                src_file_path = os.path.join(copy_img_src, file_name)
-                dst_file_path = os.path.join(copy_img_dst, file_name)
-                shutil.copy2(src_file_path, dst_file_path)
+                # Copy the sourceFrames 
+                for file_name in os.listdir(copy_img_src):
+                    src_file_path = os.path.join(copy_img_src, file_name)
+                    dst_file_path = os.path.join(copy_img_dst, file_name)
+                    shutil.copy2(src_file_path, dst_file_path)
+                pbar.update((1 / (len(src_path) * len(dataset[i]))) * 0.2)
 
 
 
 ## https://github.com/DeepLabCut/DeepLabCut/blob/main/deeplabcut/utils/conversioncode.py#L30
-def convert_csv(config_path, scorer):
+def convert_csv(config_path, scorer, pbar):
     """
         Convert (image) annotation files in folder labeled-data from csv to h5.
         This function allows the user to manually edit the csv (e.g. to correct the scorer name and then convert it into hdf format).
@@ -221,6 +271,9 @@ def convert_csv(config_path, scorer):
 
         scorer: string, optional
             If a string is given, then the scorer/annotator in all csv and hdf files that are changed, will be overwritten with this name.
+            
+        pbar : tqdm
+            Progress bar object.
 
         Examples
         --------
@@ -233,11 +286,12 @@ def convert_csv(config_path, scorer):
         --------
     """
     deeplabcut.convertcsv2h5(config_path, userfeedback=False, scorer=scorer)
+    pbar.update(1)
 
 
 
 ## https://github.com/DeepLabCut/DeepLabCut/blob/main/deeplabcut/gui/tabs/create_training_dataset.py#L84 
-def train(config_path, vdos):
+def train(config_path, vdos, pbar, destfolder):
     """
         Train a network for a project.
 
@@ -248,12 +302,21 @@ def train(config_path, vdos):
 
         vdos : list
             List of video files to use for training.
+            
+        pbar : tqdm
+            Progress bar object.
+        
+        destfolder : string or None, optional, default=None
+            Specifies the destination folder that was used for storing analysis data. 
+            If 'None', the path of the video file is used.
     """
     ## create training dataset
     deeplabcut.create_training_dataset(config_path,augmenter_type='imgaug')
-    
+    pbar.update(0.5)
+
     ## start training
     deeplabcut.train_network(config_path,shuffle=1,trainingsetindex=0,gputouse=None,max_snapshots_to_keep=5,autotune=False,displayiters=300,saveiters=1500,maxiters=30000,allow_growth=True)
+    pbar.update(0.5)
     
     # ## start evaluating
     # deeplabcut.evaluate_network(config_path,Shuffles=[1],plotting=True)
@@ -263,7 +326,7 @@ def train(config_path, vdos):
     # deeplabcut.filterpredictions(config_path,vdos)
     
     # ## create labeled video
-    # deeplabcut.create_labeled_video(config_path,vdos,draw_skeleton=True)
+    # deeplabcut.create_labeled_video(config_path,vdos, draw_skeleton=True, destfolder=destfolder)
     # deeplabcut.plot_trajectories(config_path,vdos,showfigures=True)
 
 
@@ -272,7 +335,7 @@ def train(config_path, vdos):
 
 
 ## AI-Hub의 데이터셋을 이용하여 DLC 학습하기 : frames_to_video -> create_new_project  -> config_edit -> json_to_csv -> convert_csv -> create_training_dataset -> training in CoLab
-def start(project, scorer, working_directory, src_path, lab_path, dataset):
+def start(project, scorer, working_directory, src_path, lab_path, dataset, destfolder):
     """
         Data pre-processing and sensor training.
 
@@ -286,6 +349,10 @@ def start(project, scorer, working_directory, src_path, lab_path, dataset):
 
         working_directory : string, optional
             Full path of the working directory.
+            
+        destfolder : string or None, optional, default=None
+            Where the labeled video will be stored. 
+            If 'None', the path of your project's videos folder is used.
 
         src_path : list
             Full path of the source data folder. Frame-by-frame image files, and the folder where the video files will be stored.
@@ -298,6 +365,12 @@ def start(project, scorer, working_directory, src_path, lab_path, dataset):
 
         More detailed explanation
         -------------------------
+        total_steps : int
+            Total number of steps.
+
+        pbar : tqdm
+            Progress bar object.
+
         config_path : string
             Created project config path.
 
@@ -314,29 +387,15 @@ def start(project, scorer, working_directory, src_path, lab_path, dataset):
         convert_csv( ) : Convert CSV to HDF5.
         train( ) : Training dataset.
     """
-    total_steps = 7
-    with tqdm(total=total_steps) as pbar:
-        frames_to_video(src_path, dataset)
-        pbar.update(1)
-
-        config_path, vdos = create_new(project, scorer, working_directory, src_path, dataset)
-        pbar.update(1)
-
+    total_steps = 6
+    with tqdm(total=total_steps, ncols=85) as pbar:
+        frames_to_video(src_path, dataset, pbar)
+        config_path, vdos = create_new(project, scorer, working_directory, src_path, dataset, pbar)
         conclusion_path = os.path.dirname(config_path)
-        pbar.update(1)
-
-        config_edit(config_path)
-        pbar.update(1)
-
-        json_to_csv(scorer, src_path, lab_path, conclusion_path, dataset)
-        pbar.update(1)
-
-        convert_csv(config_path, scorer)        # --> matrix 주의
-        pbar.update(1)
-
-        # colab에서 학습 진행할 경우, train() 함수는 colab에서 실행
-        train(config_path, vdos)               
-        pbar.update(1)
+        config_edit(config_path, pbar)
+        json_to_csv(scorer, src_path, lab_path, conclusion_path, dataset, pbar)
+        convert_csv(config_path, scorer, pbar)        # --> matrix 주의
+        train(config_path, vdos, pbar, destfolder)          # colab에서 학습 진행할 경우, train() 함수는 colab에서 실행
 
 
 
@@ -345,7 +404,9 @@ def main():
         Make settings before starting the programme.
 
         Variable description
-        --------------------        
+        --------------------  
+        Variables that need to be initialized
+        -------------------------------------
         project : string
             Name of the project to create.
 
@@ -357,7 +418,13 @@ def main():
 
         working_directory : string, optional
             Full path of the working directory.
-
+        
+        destfolder : string or None, optional, default=None
+            Where the labeled video will be stored. 
+            If 'None', the path of your project's videos folder is used.
+            
+        Variables that don't need to be declared
+        ----------------------------------------
         action : list
             List of the action of dataset. 
 
@@ -377,10 +444,12 @@ def main():
             Total number of the data.
     """
     project = 'preLabeled'
-    scorer = 'test01'
+    scorer = 'test01'           # project 유/무 확인하기
     data_path = "/home/dlc/DLC/_mina/data/AI-Hub/poseEstimation/Validation/DOG"
-    working_directory=r'\home\dlc\DLC\_mina\project'
-    action = ["SIT", "LYING", "BODYLOWER"]                 
+    working_directory='/home/dlc/DLC/_mina/project'
+    destfolder = '/home/dlc/DLC/_mina'
+    action = ["SIT", "BODYLOWER", "WALKRUN"]          
+           
     src_path = []                  
     lab_path = []                  
     dataset = []                 
@@ -395,7 +464,7 @@ def main():
         total += data_cnt
     print('total : ', total)
 
-    start(project, scorer, working_directory, src_path, lab_path, dataset)
+    start(project, scorer, working_directory, src_path, lab_path, dataset, destfolder=None)
 main()
